Fixes for reflection at bc with IMEX and Rosenbrock-RK
Currently the tracking is performed from the position of stage i-i to
the position in stage i with a linear displacement. hence, the
displacement between two RK stages is computed and used to scale
the current lengthPartTrajectory in order to get a better dtFrac for the
recomputation.

proper particle tracking requries to push the particle always from
position at t^n along the path, then dtfrac is correct for the first
intersection in each stage. the path is compared to the
particle path length between two positions in order to match the lenght.
dtfrac is still wrong for subsequent reflections during the same stag.

impact of this error: particle travels further than allowed
somewhere in the mappings, the dt is to large

rosenbrock-tracking requires iStages=0,2-nRKStages
imex-tracking       requires iStages=  2-nRKStages