Add mortar treatment to Tracing
With the current halo implementation, Tracing is performed on the local
sides of an element. This breaks at the latest once we try to find the
neighbor element of this side.

Loop through all small mortar sides associated with a big mortar side
and try to find the one we passed through. If no side is hit, assume we
stayed in the same element and output an warning.

The last case should not happen except for extremely curved elements,
but the current implementation is still wonky.