Insert option with numbered name ($ replaced by number), do not remove numberedMulti from list
Both are required to fix the loadbalance case. We need the numbered name with the correct value since we pretend to have it found in the parameter.ini. Furthermore, the numberedMulti entry (with the [$]) is required if we did not find the parameter in the parameter.ini and want to use the default value